Add special tie header for generated files
