+++ test/fixtures/bundlex_exs_ref_generated/cnode/example_types.h
