--- test/fixtures/bundlex_exs_ref_generated/cnode/example_types_definitions.h
@@ -1,10 +1,0 @@
-
-#pragma once
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#ifdef __cplusplus
-}
-#endif
