+++ test/fixtures/cnode_ref_generated/example_types.h
@@ -1,5 +1,5 @@
 #pragma once
 #ifdef BUNDLEX_CNODE
-#include "cnode/example.h"
+
 #include "cnode/example_types.h"
 #endif
