+++ test/fixtures/nif_ref_generated/example_types.h
@@ -1,5 +1,5 @@
 #pragma once
 #ifdef BUNDLEX_NIF
-#include "nif/example.h"
+
 #include "nif/example_types.h"
 #endif
