+++ test/fixtures/unified_ref_generated/cnode/example_types.h
