+++ test/fixtures/unified_ref_generated/example.h
@@ -1,12 +1,10 @@
 #pragma once
-
 #ifdef BUNDLEX_NIF
 #include "nif/example.h"
-#include "nif/example_types_definitions.h"
+#include "nif/example_types.h"
 #endif
 
 #ifdef BUNDLEX_CNODE
 #include "cnode/example.h"
-#include "cnode/example_types_definitions.h"
+#include "cnode/example_types.h"
 #endif
-
