+++ test/fixtures/unified_ref_generated/example_types.h
@@ -1,5 +1,10 @@
 #pragma once
+#ifdef BUNDLEX_NIF
+
+#include "nif/example_types.h"
+#endif
+
 #ifdef BUNDLEX_CNODE
-#include "cnode/example.h"
+
 #include "cnode/example_types.h"
 #endif
