+++ test/fixtures/unified_ref_generated/nif/example_types.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <erl_nif.h>
+#include <stdint.h>
+#include <stdio.h>
+#include <unifex/payload.h>
+#include <unifex/unifex.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#ifdef __cplusplus
+}
+#endif
